@@ -12,13 +12,14 @@
     "deploy": "npm run pages:build && wrangler pages deploy"
   },
   "dependencies": {
-    "@hookform/resolvers": "3.9.0",
+    "@hookform/resolvers": "3.9.1",
     "@radix-ui/react-accordion": "1.2.1",
+    "@radix-ui/react-alert-dialog": "1.1.2",
     "@radix-ui/react-avatar": "1.1.1",
     "@radix-ui/react-collapsible": "1.1.1",
     "@radix-ui/react-dialog": "1.1.2",
     "@radix-ui/react-dropdown-menu": "2.1.2",
-    "@radix-ui/react-icons": "1.3.0",
+    "@radix-ui/react-icons": "1.3.1",
     "@radix-ui/react-label": "2.1.0",
     "@radix-ui/react-menubar": "1.1.2",
     "@radix-ui/react-navigation-menu": "1.2.1",
@@ -28,40 +29,47 @@
     "@radix-ui/react-separator": "1.1.0",
     "@radix-ui/react-slot": "1.1.0",
     "@radix-ui/react-tabs": "1.1.1",
-    "@radix-ui/react-toast": "1.2.2",
+    "@radix-ui/react-toast": "^1.2.2",
     "@radix-ui/react-tooltip": "1.1.3",
     "@tanstack/react-table": "8.20.5",
     "class-variance-authority": "0.7.0",
     "clsx": "2.1.1",
     "gravatar-url": "4.0.1",
-    "lucide-react": "0.453.0",
-    "nanoid": "5.0.7",
-    "next": "14.2.16",
+    "lucide-react": "0.456.0",
+    "nanoid": "5.0.8",
+    "next": "15.0.3",
     "next-auth": "5.0.0-beta.25",
-    "next-plausible": "3.12.2",
+    "next-plausible": "3.12.4",
     "next-seo": "6.6.0",
-    "oauth4webapi": "2.17.0",
     "react": "18",
     "react-dom": "18",
-    "react-hook-form": "7.53.1",
+    "react-hook-form": "7.53.2",
     "tailwind-merge": "2.5.4",
     "tailwindcss-animate": "1.0.7",
+    "tailwindcss-react-aria-components": "1.1.6",
     "zod": "3.23.8"
   },
   "devDependencies": {
     "@cloudflare/next-on-pages": "1.13.5",
-    "@cloudflare/workers-types": "4.20241022.0",
+    "@cloudflare/workers-types": "4.20241106.0",
     "@types/node": "22",
     "@types/react": "18",
     "@types/react-dom": "18",
     "autoprefixer": "10.4.20",
     "eslint": "8",
-    "eslint-config-next": "14.2.16",
+    "eslint-config-next": "15.0.3",
+    "eslint-config-prettier": "9.1.0",
     "eslint-plugin-next-on-pages": "1.13.5",
     "postcss": "8.4.47",
     "tailwindcss": "3.4.14",
     "typescript": "5",
-    "vercel": "37.12.1",
-    "wrangler": "3.83.0"
+    "vercel": "37.14.0",
+    "wrangler": "3.86.0"
+  },
+  "prettier": {
+    "arrowParens": "always",
+    "singleQuote": true,
+    "tabWidth": 2,
+    "trailingComma": "none"
   }
 }
