@@ -18,35 +18,13 @@ import { Card, CardContent, CardHeader, CardTitle } from '@/components/ui/card'
 import { Collapsible, CollapsibleContent, CollapsibleTrigger } from '@/components/ui/collapsible'
 import { Form, FormControl, FormDescription, FormField, FormItem, FormLabel, FormMessage } from '@/components/ui/form'
 import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from '@/components/ui/select'
-import { ApplicationTypes, GrantTypes, tokenAuthenticationMethods, TokenEndpointAuthMethods } from '@/lib/consts'
+import { ApplicationTypes, tokenAuthenticationMethods } from '@/lib/consts'
 import { urlEncode } from '@/lib/url'
 import { getTemplateById } from '@/lib/templates'
 
+import { clientSchema as createClientSchema } from '../schema'
+
 const localStorageItem = 'client'
-
-const urlSchema = z.union([
-  z.string().trim().url().startsWith('http://'),
-  z.string().trim().url().startsWith('https://'),
-], { errorMap: () => ({ message: 'URL must begin with http:// or https://' }) })
-const createClientSchema = z.object({
-  template: z.string().optional(),
-  applicationType: z.enum(Object.values(ApplicationTypes) as [ApplicationType]),
-  grantTypes: z.array(z.enum(Object.values(GrantTypes) as [GrantType])),
-  tokenEndpointAuthMethod: z.enum(Object.values(TokenEndpointAuthMethods) as [TokenEndpointAuthMethod]),
-
-  name: z.string(),
-  uri: z.union([z.literal(''), urlSchema]).optional(),
-  allowedCorsOrigins: z.array(z.string()).optional(),
-  scopes: z.array(z.string()).optional(),
-  audiences: z.array(z.string()).optional(),
-  redirectUris: urlSchema.array().min(1, { message: 'At least one redirect URI is required' }),
-  postLogoutRedirectUris: urlSchema.array().optional(),
-
-  contacts: z.array(z.string()).optional(),
-  policyUri: z.union([z.literal(''), urlSchema]).optional(),
-  tosUri: z.union([z.literal(''), urlSchema]).optional(),
-  logoUri: z.union([z.literal(''), urlSchema]).optional(),
-})
 
 export const runtime = 'edge'
 export const dynamic = 'force-static'
