@@ -1,11 +1,10 @@
 "use client"
 
 import * as React from "react"
-import { Cross2Icon } from "@radix-ui/react-icons"
 import * as ToastPrimitives from "@radix-ui/react-toast"
 import { cva, type VariantProps } from "class-variance-authority"
-
 import { cn } from "@/lib/utils"
+import { Cross2Icon } from "@radix-ui/react-icons"
 
 const ToastProvider = ToastPrimitives.Provider
 
